+++ noodles-sam/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+  * sam/record: Add mutable getter for read name (`Record::read_name_mut`).
+
 ## 0.4.0 - 2021-09-23
 
 ### Changed