+++ noodles-sam/CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Added
 
   * sam/record: Add mutable getters for flags (`Record::flags_mut`), read name
-    (`Record::read_name_mut`), and mapping quality
-    (`Record::mapping_quality_mut`).
+    (`Record::read_name_mut`), mapping quality (`Record::mapping_quality_mut`),
+    and template length (`Record::template_length_mut`).
 
 ## 0.4.0 - 2021-09-23