+++ noodles-sam/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Added
 
-  * sam/record: Add mutable getter for reference sequence name
-    (`Record::reference_sequence_name_mut`).
+  * sam/record: Add mutable getter for position (`Record::position_mut`) and
+    reference sequence name (`Record::reference_sequence_name_mut`).
 
 ## 0.5.0 - 2021-10-01