+++ noodles-sam/CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Added
 
-  * sam/record: Add mutable getters for position (`Record::position_mut`),
-    mate reference sequence name (`Record::mate_reference_sequence_name_mut`),
-    and reference sequence name (`Record::reference_sequence_name_mut`).
+  * sam/record: Add mutable getters for position (`Record::position_mut`), mate
+    position (`Record::mate_position_mut`), mate reference sequence name
+    (`Record::mate_reference_sequence_name_mut`), and reference sequence name
+    (`Record::reference_sequence_name_mut`).
 
 ## 0.5.0 - 2021-10-01